Own component to show entries on landing page
