+++ src/utils/dataTestIds.ts
@@ -307,9 +307,7 @@ export const dataTestId = {
     keywords: 'public-registration-keywords',
     license: 'public-registration-license',
     noLinkOrFilesWarning: 'no-link-or-files-warning-typhography',
-    npi: 'public-registration-npi',
     openFileButton: 'open-file-button',
-    primaryLanguage: 'public-registration-primary-language',
     projectsAccordion: 'projects-accordion',
     projectTitle: 'project-title',
     publicationsUsingDatasetAccordion: 'publications-using-dataset-accordion',
